HW 15: Update integration tests
